Change bibtex style and moved hyperref
The bibtex style has been changed from siam to ieeetr.
There was some attempt to stop the bibliography replacing repeated authors with "----" from this link: https://tex.stackexchange.com/questions/29381/is-it-normal-for-bibtex-to-replace-similar-author-names-with

Hyperref was moved and Bookmarks added to try and stop needing to compile twice from this link: https://tex.stackexchange.com/questions/30330/how-can-i-avoid-compiling-twice

Both problems need further investigating.