Adding detail, plot list colours and appendices
Added plots and detail to chapters 3 and 4.
The colours.tex file defined the colour lists used in pgfplots. This will need updating if future plots need a different number of colours
An appendix with the errors and gains from the MDI experiment has been added.
The table separation was also changed to spread the table columns and rows out a bit.